Remove useless cast breaking build on 32-bits os

+++ src/providers/ogr/qgsogrprovider.cpp
@@ -583,7 +583,7 @@ QString QgsOgrProvider::ogrWkbGeometryTypeName( OGRwkbGeometryType type ) const
     return geom;
   }
 
-  switch ( ( long )type )
+  switch ( type )
   {
     case wkbUnknown:
       geom = QStringLiteral( "Unknown" );
